--- alembic/versions/20250408_070731_77c785b4d309_default_migration_message.py
@@ -1,87 +1,0 @@
-"""Default migration message
-
-Revision ID: 77c785b4d309
-Revises: b05c03401442
-Create Date: 2025-04-08 07:07:31.862512
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic_postgresql_enum import TableReference
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "77c785b4d309"
-down_revision: str | None = "b05c03401442"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("ion_chanel_model", sa.Column("emodel_id", sa.Uuid(), nullable=False))
-    op.create_foreign_key(
-        op.f("fk_ion_chanel_model_emodel_id_emodel"),
-        "ion_chanel_model",
-        "emodel",
-        ["emodel_id"],
-        ["id"],
-    )
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="entitytype",
-        new_values=[
-            "analysis_software_source_code",
-            "emodel",
-            "experimental_bouton_density",
-            "experimental_neuron_density",
-            "experimental_synapses_per_connection",
-            "memodel",
-            "mesh",
-            "reconstruction_morphology",
-            "single_cell_experimental_trace",
-            "single_neuron_simulation",
-            "single_neuron_synaptome",
-            "single_neuron_synaptome_simulation",
-            "ion_channel_model",
-        ],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="entity", column_name="type")
-        ],
-        enum_values_to_rename=[],
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.sync_enum_values(
-        enum_schema="public",
-        enum_name="entitytype",
-        new_values=[
-            "analysis_software_source_code",
-            "emodel",
-            "experimental_bouton_density",
-            "experimental_neuron_density",
-            "experimental_synapses_per_connection",
-            "memodel",
-            "mesh",
-            "reconstruction_morphology",
-            "single_cell_experimental_trace",
-            "single_neuron_simulation",
-            "single_neuron_synaptome",
-            "single_neuron_synaptome_simulation",
-        ],
-        affected_columns=[
-            TableReference(table_schema="public", table_name="entity", column_name="type")
-        ],
-        enum_values_to_rename=[],
-    )
-    op.drop_constraint(
-        op.f("fk_ion_chanel_model_emodel_id_emodel"), "ion_chanel_model", type_="foreignkey"
-    )
-    op.drop_column("ion_chanel_model", "emodel_id")
-    # ### end Alembic commands ###
