+++ alembic/versions/20250409_035752_bedbe97c0c36_default_migration_message.py
@@ -1,0 +1,185 @@
+"""Default migration message
+
+Revision ID: bedbe97c0c36
+Revises: 72c5253bf9b1
+Create Date: 2025-04-09 03:57:52.170547
+
+"""
+
+from collections.abc import Sequence
+
+import sqlalchemy as sa
+from alembic_postgresql_enum import TableReference
+from sqlalchemy.dialects import postgresql
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "bedbe97c0c36"
+down_revision: str | None = "72c5253bf9b1"
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "ion",
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column("name", sa.String(), nullable=False),
+        sa.Column(
+            "creation_date",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+        sa.Column(
+            "update_date",
+            sa.DateTime(timezone=True),
+            server_default=sa.text("now()"),
+            nullable=False,
+        ),
+        sa.PrimaryKeyConstraint("id", name=op.f("pk_ion")),
+    )
+    op.create_index(op.f("ix_ion_creation_date"), "ion", ["creation_date"], unique=False)
+    op.create_index(op.f("ix_ion_name"), "ion", ["name"], unique=True)
+    op.create_table(
+        "ion_channel_model",
+        sa.Column("id", sa.Uuid(), nullable=False),
+        sa.Column("name", sa.String(), nullable=False),
+        sa.Column("description", sa.String(), nullable=False),
+        sa.Column("identifier", sa.String(), nullable=False),
+        sa.Column("modelId", sa.String(), nullable=False),
+        sa.Column("is_ljp_corrected", sa.Boolean(), nullable=False),
+        sa.Column("is_temperature_dependent", sa.Boolean(), nullable=False),
+        sa.Column("temperature_celsius", sa.Integer(), nullable=False),
+        sa.Column("nmodl_parameters", postgresql.JSONB(astext_type=sa.Text()), nullable=False),
+        sa.Column("emodel_id", sa.Uuid(), nullable=False),
+        sa.Column("description_vector", postgresql.TSVECTOR(), nullable=True),
+        sa.Column("brain_region_id", sa.BigInteger(), nullable=False),
+        sa.Column("species_id", sa.Uuid(), nullable=False),
+        sa.Column("strain_id", sa.Uuid(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["brain_region_id"],
+            ["brain_region.id"],
+            name=op.f("fk_ion_channel_model_brain_region_id_brain_region"),
+        ),
+        sa.ForeignKeyConstraint(
+            ["emodel_id"], ["emodel.id"], name=op.f("fk_ion_channel_model_emodel_id_emodel")
+        ),
+        sa.ForeignKeyConstraint(["id"], ["entity.id"], name=op.f("fk_ion_channel_model_id_entity")),
+        sa.ForeignKeyConstraint(
+            ["species_id"], ["species.id"], name=op.f("fk_ion_channel_model_species_id_species")
+        ),
+        sa.ForeignKeyConstraint(
+            ["strain_id", "species_id"],
+            ["strain.id", "strain.species_id"],
+            name="fk_ion_channel_model_strain_id_species_id",
+        ),
+        sa.PrimaryKeyConstraint("id", name=op.f("pk_ion_channel_model")),
+    )
+    op.create_index(
+        op.f("ix_ion_channel_model_brain_region_id"),
+        "ion_channel_model",
+        ["brain_region_id"],
+        unique=False,
+    )
+    op.create_index(
+        "ix_ion_channel_model_description_vector",
+        "ion_channel_model",
+        ["description_vector"],
+        unique=False,
+        postgresql_using="gin",
+    )
+    op.create_index(op.f("ix_ion_channel_model_name"), "ion_channel_model", ["name"], unique=False)
+    op.create_index(
+        op.f("ix_ion_channel_model_species_id"), "ion_channel_model", ["species_id"], unique=False
+    )
+    op.create_index(
+        op.f("ix_ion_channel_model_strain_id"), "ion_channel_model", ["strain_id"], unique=False
+    )
+    op.create_table(
+        "ion_channel_association",
+        sa.Column("ion_id", sa.Uuid(), nullable=False),
+        sa.Column("ion_channel_model_id", sa.Uuid(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["ion_channel_model_id"],
+            ["ion_channel_model.id"],
+            name=op.f("fk_ion_channel_association_ion_channel_model_id_ion_channel_model"),
+            ondelete="CASCADE",
+        ),
+        sa.ForeignKeyConstraint(
+            ["ion_id"],
+            ["ion.id"],
+            name=op.f("fk_ion_channel_association_ion_id_ion"),
+            ondelete="CASCADE",
+        ),
+        sa.PrimaryKeyConstraint(
+            "ion_id", "ion_channel_model_id", name=op.f("pk_ion_channel_association")
+        ),
+    )
+    op.sync_enum_values(
+        enum_schema="public",
+        enum_name="entitytype",
+        new_values=[
+            "analysis_software_source_code",
+            "emodel",
+            "experimental_bouton_density",
+            "experimental_neuron_density",
+            "experimental_synapses_per_connection",
+            "memodel",
+            "mesh",
+            "reconstruction_morphology",
+            "single_cell_experimental_trace",
+            "single_neuron_simulation",
+            "single_neuron_synaptome",
+            "single_neuron_synaptome_simulation",
+            "ion_channel_model",
+        ],
+        affected_columns=[
+            TableReference(table_schema="public", table_name="entity", column_name="type")
+        ],
+        enum_values_to_rename=[],
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.sync_enum_values(
+        enum_schema="public",
+        enum_name="entitytype",
+        new_values=[
+            "analysis_software_source_code",
+            "emodel",
+            "experimental_bouton_density",
+            "experimental_neuron_density",
+            "experimental_synapses_per_connection",
+            "memodel",
+            "mesh",
+            "reconstruction_morphology",
+            "single_cell_experimental_trace",
+            "single_neuron_simulation",
+            "single_neuron_synaptome",
+            "single_neuron_synaptome_simulation",
+        ],
+        affected_columns=[
+            TableReference(table_schema="public", table_name="entity", column_name="type")
+        ],
+        enum_values_to_rename=[],
+    )
+    op.drop_table("ion_channel_association")
+    op.drop_index(op.f("ix_ion_channel_model_strain_id"), table_name="ion_channel_model")
+    op.drop_index(op.f("ix_ion_channel_model_species_id"), table_name="ion_channel_model")
+    op.drop_index(op.f("ix_ion_channel_model_name"), table_name="ion_channel_model")
+    op.drop_index(
+        "ix_ion_channel_model_description_vector",
+        table_name="ion_channel_model",
+        postgresql_using="gin",
+    )
+    op.drop_index(op.f("ix_ion_channel_model_brain_region_id"), table_name="ion_channel_model")
+    op.drop_table("ion_channel_model")
+    op.drop_index(op.f("ix_ion_name"), table_name="ion")
+    op.drop_index(op.f("ix_ion_creation_date"), table_name="ion")
+    op.drop_table("ion")
+    # ### end Alembic commands ###
